Add `--reuse-first-build` flag to `runtests.py`

+++ harness/machine_types/base_machine.py
@@ -271,16 +271,25 @@ class BaseMachine(metaclass=ABCMeta):
             The exit status of the build command.
 
         """
-        message = f"Start of buiding executable.\n"
-        self.logger.doInfoLogging(message)
+        self.logger.doInfoLogging("Start of buiding executable.")
 
         currentdir = os.getcwd()
-        message = f"The initial directory is {currentdir}"
-        self.logger.doInfoLogging(message)
+        self.logger.doInfoLogging(f"The initial directory is {currentdir}")
 
+        path_to_source = self.apptest.get_path_to_source()
+        path_to_test_source = self.apptest.get_path_to_test_source()
         path_to_build_directory = self.apptest.get_path_to_workspace_build()
-        message = f"The build directory is {path_to_build_directory}"
-        self.logger.doInfoLogging(message)
+        path_to_runarchive_directory = self.apptest.get_path_to_runarchive()
+
+        # Use Error threshold to show these messages all the time
+        self.logger.doErrorLogging(f"Path to Source: {path_to_source}")
+        self.logger.doErrorLogging(f"Path to Build: {path_to_build_directory}")
+        self.logger.doErrorLogging(f"Path to Run_Archive: {path_to_runarchive_directory}")
+
+        if 'RGT_REUSE_BUILD_FROM' in os.environ and \
+                os.path.exists(os.environ['RGT_REUSE_BUILD_FROM']):
+            self.logger.doInfoLogging(f"Skipping build, re-using the build from {os.environ['RGT_REUSE_BUILD_FROM']}")
+            return 0
 
         # Copy the source to the build directory.
         copy_rc = self._copy_source_to_build_directory()
@@ -288,16 +297,14 @@ class BaseMachine(metaclass=ABCMeta):
         if not copy_rc == 0:
             return copy_rc
 
-        message = f"Copied source to build directory.\n"
-        self.logger.doInfoLogging(message)
+        self.logger.doInfoLogging(f"Copied source to build directory.")
 
         # Get the environment using the build runtime environment file.
         new_env = None
         filename = self.build_runtime_environment_command_file
 
         if filename != "":
-            message = f"The build runtime environmental file is {filename}."
-            self.logger.doInfoLogging(message)
+            self.logger.doInfoLogging(f"The build runtime environmental file is {filename}.")
             new_env = linux_utilities.get_new_environment(self,filename)
             message = f"The new build environment is as follows:\n"
             message += str(new_env)
@@ -429,21 +436,16 @@ class BaseMachine(metaclass=ABCMeta):
     def _copy_source_to_build_directory(self):
         path_to_source = self.apptest.get_path_to_source()
         path_to_test_source = self.apptest.get_path_to_test_source()
-        # Use Error threshold to show this message all the time
-        self.logger.doErrorLogging(f"Path to Source: {path_to_source}")
-
         path_to_build_directory = self.apptest.get_path_to_workspace_build()
-        # Use Error threshold to show this message all the time
-        self.logger.doErrorLogging(f"Path to Build: {path_to_build_directory}")
 
-        path_to_runarchive_directory = self.apptest.get_path_to_runarchive()
-        # Use Error threshold to show this message all the time
-        self.logger.doErrorLogging(f"Path to Run_Archive: {path_to_runarchive_directory}")
+        if 'RGT_REUSE_BUILD_FROM' in os.environ and \
+                os.path.exists(os.environ['RGT_REUSE_BUILD_FROM']):
+            self.logger.doInfoLogging("RGT_REUSE_BUILD_FROM set, skipping copying Source.")
+            return 0
 
         shutil.copytree(src=path_to_source,
                         dst=path_to_build_directory,
                         symlinks=True)
-
         # If a Source directory exists inside test, overlay that over source directory
         if os.path.exists(path_to_test_source):
             # Python 3.8 adds the dirs_exist_ok keyword to allow overwriting a destination
